requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
werkzeug==3.0.1
ffmpeg-python==0.2.0

# Document conversion dependencies
python-docx==1.1.0
pypdf==3.17.0
pdf2docx==0.5.8
openpyxl==3.1.2
reportlab==4.0.7
markdown==3.5.1

# OCR dependencies
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3

